feat: initial lsptrace implementation (named pipes)
A very barebones lsptrace implementation that supports named pipes (or
rather, right now just unix domain sockets, which is what roslyn seems
to use on unix/osx). It would have been simpler to just do stdin/stdout
and that's what I started with (dummyclient and dummyserver were testing
this) but the only reason I'm writing this is to help debug roslyn and
rzls not for lsp in general, so I hacked in the more complicated named
pipes support.

Currently using the Dog's Breakfast Style coding pattern.

Usage:
- build lsptrace binary ./build.sh
- point to bin/lsptrace binary in language server "cmd" configuration
  e.g. instead of `cmd = "dotnet ..."` in roslyn.nvim, it becomes `cmd =
  "lsptrace dotnet ..."`
- see output in debug.log. Currently hardcoded, and currently everything
  is being written there (actual debug logs and the lsptraces in
  question). Writing log to file is a necessity for this program because
  of its nature (being called as a child subprocess in another
  application in a context where stdin/stdout are already being used for
  another process i.e. in this case, negotiating the named pipe
  transport initialization)

Next:
- start actually writing out .lsptrace files, splitting from debug.log:
  in this process, should start parsing the jsonrpc messages and writing
  out the lsptrace format instead of the raw jsonrpc e.g. timestamp and
  serverName
- once .lsptrace files are good, basic parser and ui
- don't hardcode debug.log path
- consider moving off Dog's Breakfast Style.