fix: response method properly populated from corresponding request
LSP response messages don't include the method because it already
includes "id" which can be mapped to a corresponding request. This is
useful for actually running the protocol, but for tracing it's useful to
populate the method as we're tracing so we don't have to implement the
same logic on viewing the trace file.

I tried previously to do this directly in LSPTracer but it didn't work
because we have a separate tracer for client and server streams. Had to
create a sync map for request which is shared between the tracers.