cleanup: use In/Out pattern for interceptor and create channelwriter
channel writer will sit in front of the input stream and stream to a
channel.

Interceptor will now use an input channel of []byte instead of acting as
a Writer. It was a little strange for this to be Writer in the first
place, and it feels more consistent to use channels for in/out in the
pipeline.

These changes are setting the stage for a Pipeline which will represent
the pipeline from lsp input stream to the lsp output stream + lsp trace
output stream.