feat: lsptrace is configurable from cmdline and env vars
if we can construct the lsptrace call flexibly like in roslyn.nvim
configuration then we can pass in flags to configure the behavior. if we
can't construct the call so flexibly like in vscode-sharp, then we need
to use environment variables (easiest way is to set env vars when
calling the vscode exe from commandline).

the code in the main loop was cleaned up to parse flags and environment
variables. also refactored the creation of the clientIO and serverIO
pipes for the pipeline, so lsptrace can support roslyn's named pipe flow
as well as the stdio flow.